\name{ProbKMAcpp-package}
\alias{ProbKMAcpp-package}
\alias{ProbKMAcpp}
\docType{package}
\title{\packageTitle{ProbKMAcpp}}
\description{\packageDescription{ProbKMAcpp}}
\section{Package Content}{\packageIndices{ProbKMAcpp}}
\author{\packageAuthor{ProbKMAcpp}}
\section{Maintainer}{\packageMaintainer{ProbKMAcpp}}
\keyword{package}
